backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
google-genai==1.7.0
gunicorn==21.2.0
werkzeug==2.2.3
vertexai==1.71.1